requirements.txt
pandas==2.2.3
openpyxl==3.1.5
streamlit==1.43.1
bs4==0.0.2
pyyaml==6.0.2